# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  available: 'available',
  reserved: 'reserved',
  sold: 'sold',
} as const;

export interface Listing {
  id: number;
  farmerName: string;
  crop: string;
  quantity: number;
  unit: string;
  askingPrice: number;
  marketPrice: number;
  location: string;
  harvestDate: string;
  status: ListingStatus;
  verified: boolean;
  createdAt?: string;
}

export interface ListingInput {
  /** @minLength 1 */
  farmerName: string;
  /** @minLength 1 */
  crop: string;
  /** @minimum 0.1 */
  quantity: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  askingPrice: number;
  /** @minimum 0 */
  marketPrice: number;
  /** @minLength 1 */
  location: string;
  harvestDate: string;
}

export type ListingUpdateStatus = typeof ListingUpdateStatus[keyof typeof ListingUpdateStatus];


export const ListingUpdateStatus = {
  available: 'available',
  reserved: 'reserved',
  sold: 'sold',
} as const;

export interface ListingUpdate {
  /** @minimum 0 */
  askingPrice?: number;
  status?: ListingUpdateStatus;
  /** @minimum 0.1 */
  quantity?: number;
}

export interface Buyer {
  id: number;
  name: string;
  category: string;
  location: string;
  crops: string[];
  verified: boolean;
  responseTime: string;
  rating?: number;
}

export interface PriceSnapshot {
  crop: string;
  localMandi: number;
  fairPrice: number;
  bestBuyerPrice: number;
  unit: string;
  change: number;
}

export interface TransportRequestInput {
  listingId: number;
  /** @minLength 1 */
  pickupLocation: string;
  pickupDate: string;
  /** @minLength 1 */
  vehicleType: string;
}

export type TransportRequestStatus = typeof TransportRequestStatus[keyof typeof TransportRequestStatus];


export const TransportRequestStatus = {
  requested: 'requested',
  assigned: 'assigned',
  completed: 'completed',
} as const;

export interface TransportRequest {
  id: number;
  listingId: number;
  pickupLocation: string;
  pickupDate: string;
  vehicleType: string;
  status: TransportRequestStatus;
  estimatedCost: number;
  createdAt?: string;
}

export interface DashboardSummary {
  farmersConnected: number;
  fairTradeValue: number;
  availableLots: number;
  verifiedBuyers: number;
  averagePriceLift: number;
}

export type ListListingsParams = {
crop?: string;
location?: string;
status?: ListListingsStatus;
};

export type ListListingsStatus = typeof ListListingsStatus[keyof typeof ListListingsStatus];


export const ListListingsStatus = {
  available: 'available',
  reserved: 'reserved',
  sold: 'sold',
} as const;