# lib/api-zod/src/generated/types/buyer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Buyer {
  id: number;
  name: string;
  category: string;
  location: string;
  crops: string[];
  verified: boolean;
  responseTime: string;
  rating?: number;
}