# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  farmersConnected: number;
  fairTradeValue: number;
  availableLots: number;
  verifiedBuyers: number;
  averagePriceLift: number;
}