# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ListingStatus } from './listingStatus';

export interface Listing {
  id: number;
  farmerName: string;
  crop: string;
  quantity: number;
  unit: string;
  askingPrice: number;
  marketPrice: number;
  location: string;
  harvestDate: Date;
  status: ListingStatus;
  verified: boolean;
  createdAt?: Date;
}