# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ListingInput {
  /** @minLength 1 */
  farmerName: string;
  /** @minLength 1 */
  crop: string;
  /** @minimum 0.1 */
  quantity: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  askingPrice: number;
  /** @minimum 0 */
  marketPrice: number;
  /** @minLength 1 */
  location: string;
  harvestDate: Date;
}