# lib/api-zod/src/generated/types/listingUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ListingUpdateStatus } from './listingUpdateStatus';

export interface ListingUpdate {
  /** @minimum 0 */
  askingPrice?: number;
  status?: ListingUpdateStatus;
  /** @minimum 0.1 */
  quantity?: number;
}