# lib/api-zod/src/generated/types/priceSnapshot.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriceSnapshot {
  crop: string;
  localMandi: number;
  fairPrice: number;
  bestBuyerPrice: number;
  unit: string;
  change: number;
}