# lib/api-zod/src/generated/types/transportRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransportRequestStatus } from './transportRequestStatus';

export interface TransportRequest {
  id: number;
  listingId: number;
  pickupLocation: string;
  pickupDate: Date;
  vehicleType: string;
  status: TransportRequestStatus;
  estimatedCost: number;
  createdAt?: Date;
}