# lib/api-zod/src/generated/types/transportRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TransportRequestInput {
  listingId: number;
  /** @minLength 1 */
  pickupLocation: string;
  pickupDate: Date;
  /** @minLength 1 */
  vehicleType: string;
}